requirements.txt
# Core data processing
pandas>=2.0
numpy>=1.24

# Excel I/O
openpyxl>=3.1
xlsxwriter>=3.1

# Google API
google-api-python-client>=2.100
google-auth>=2.23
google-auth-oauthlib>=1.1

# Database
sqlalchemy>=2.0
psycopg2-binary>=2.9

# Config
pyyaml>=6.0

# Date utilities
python-dateutil>=2.8
pytz>=2023.3

# XML parsing (for FTP response validation)
lxml>=4.9

# IPython (required by Google module)
ipython>=8.0

# Testing
pytest>=7.4
